Update 2016-09-23-Implementing-cancellable-optimistic-updates-in-Angular2.md
great work! I really like how you explained everything very clearly and to the point! Even for someone like myself, who hasn't much clue about redux, I could still follow the post and actually get everything that was in it. I LOVE the fact that you have images, makes everything so much easier to understand. Code is clear and readable (thumbsup)

I think my only comment, or question that I have after reading your article is "why would I want to do this"? (again, pls keep in mind that I'm new to this and I might be totally off). "What is my added value from doing things this way rather than how it was explained in the beginning of the article?" My first feeling is that updating the store after getting a response from our HTTP call seems pretty straightforward as compared to what feels the complexity introduced by the second approach. From what I read, the benefit is that the UI is immediately updated, which is indeed a good thing. But is that the only benefit? And what if the server returns an error? Does that mean that the UI will change back and forth?

I hope these are not dummy questions, and I apologise if they are, but this is what I am kind of missing from this post: a bit more on the "why" you'd want to do things that way and the benefits that come with that.

I hope this helps! Thank you for letting me read this. Really like your posts <3